requirements.txt
streamlit==1.36.0
pandas==2.2.2
matplotlib==3.8.4
pyyaml==6.0.1
